requirements.txt
Django==1.10.4
django-bootstrap==0.2.4
django-tinymce==2.4.0
psycopg2==2.6.2
pypiwin32==219
